+++ lib/macro_util.dart
@@ -1,3 +1,3 @@
 export 'src/declaration_builder.dart';
 export 'src/field_introspection_data.dart';
-export 'src/list.dart';
+export 'src/iterable.dart';