+++ lib/src/declaration_builder.dart
@@ -37,4 +37,53 @@ extension MacroUtilDeclarationBuilderExtension on DeclarationBuilder {
 
     return typeDecl;
   }
+
+  /// Returns the [ClassDeclaration] for the superclass of [clazz]
+  /// unless it's [Object] explicitly or implicitly.
+  Future<ClassDeclaration?> nonObjectSuperclassDeclarationOf(
+    ClassDeclaration clazz,
+  ) async {
+    final superclass = clazz.superclass;
+
+    if (superclass == null) {
+      return null;
+    }
+
+    final superType =
+        await resolve(NamedTypeAnnotationCode(name: superclass.identifier));
+
+    final objectType = await _getObjectType();
+    final isExtendingObject = await superType.isExactly(objectType);
+
+    if (isExtendingObject) {
+      return null;
+    }
+
+    final result = await unaliasedTypeDeclarationOf(superclass);
+
+    if (result is! ClassDeclaration) {
+      report(
+        Diagnostic(
+          DiagnosticMessage(
+            'Cannot resolve the superclass.',
+            target: clazz.asDiagnosticTarget,
+          ),
+          Severity.error,
+        ),
+      );
+
+      throw Exception('Cannot resolve the superclass.');
+    }
+
+    return result;
+  }
+
+  Future<StaticType> _getObjectType() async {
+    return resolve(
+      NamedTypeAnnotationCode(
+        // ignore: deprecated_member_use
+        name: await resolveIdentifier(Uri.parse('dart:core'), 'Object'),
+      ),
+    );
+  }
 }