+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0-2.dev
+
+* Added `DeclarationBuilder.nonObjectSuperclassDeclarationOf`.
+* Changed `MacroUtilListExtension` to `MacroUtilIterableExtension`.
+
 ## 0.1.0-1.dev
 
 * Initial release.